+++ packages/core-data/src/entities.js
@@ -329,7 +329,10 @@ async function loadPostTypeEntities() {
 				}/${ parentId }/revisions${
 					revisionId ? '/' + revisionId : ''
 				}`,
-			revisionKey: DEFAULT_ENTITY_KEY,
+			revisionKey:
+				isTemplate && ! window?.__experimentalTemplateActivate
+					? 'wp_id'
+					: DEFAULT_ENTITY_KEY,
 		};
 
 		if ( window.__experimentalEnableSync ) {